# Conflicts:
#	lib/Backend/IR.cpp
#	lib/Backend/IRBuilderAsmJs.cpp
#	lib/Backend/Lower.cpp
#	lib/Backend/Lower.h
#	lib/Backend/LowerMDShared.cpp
#	lib/Backend/LowerMDShared.h
#	lib/Backend/Opnd.cpp
#	lib/Backend/Opnd.h
#	lib/Backend/arm/LowerMD.h
#	lib/Backend/arm64/LowerMD.h
#	lib/Backend/i386/EncoderMD.cpp
#	lib/Backend/i386/LowererMDArch.cpp
#	lib/Runtime/ByteCode/AsmJsByteCodeDumper.cpp
#	lib/Runtime/Language/InterpreterStackFrame.h
#	lib/Runtime/Library/WasmLibrary.cpp
#	lib/WasmReader/WasmBinaryReader.h
#	lib/WasmReader/WasmByteCodeGenerator.cpp
#	lib/WasmReader/WasmModule.h

+++ lib/Backend/IntBounds.cpp
@@ -304,7 +304,8 @@ bool IntBounds::IsGreaterThanOrEqualTo(const int constantValue, const int consta
     if(offset == 1)
         return constantValue > constantBoundBase;
 
-    const int constantBound = constantBoundBase + offset;
+    // use unsigned to avoid signed int overflow
+    const int constantBound = (unsigned)constantBoundBase + (unsigned)offset;
     return
         offset >= 0
             ? constantBound >= constantBoundBase && constantValue >= constantBound
@@ -318,7 +319,8 @@ bool IntBounds::IsLessThanOrEqualTo(const int constantValue, const int constantB
     if(offset == -1)
         return constantValue < constantBoundBase;
 
-    const int constantBound = constantBoundBase + offset;
+    // use unsigned to avoid signed int overflow
+    const int constantBound = (unsigned)constantBoundBase + (unsigned)offset;
     return
         offset >= 0
             ? constantBound < constantBoundBase || constantValue <= constantBound
